test(core): merge addAtRefsType test into Spec.test.ts
Consolidates the type-level refs inference test alongside the existing
isSpec test so all Spec tests live in a single file.
